Moved config tests in separate file
